+++ lib/presentation/home/screens/home_screen.dart
@@ -5,7 +5,10 @@ import 'package:ondo/core/design_system/app_text_styles.dart';
 import 'package:ondo/core/ui/base/base_scaffold.dart';
 import 'package:ondo/presentation/home/screens/home_search_detail_screen.dart';
 import 'package:ondo/presentation/home/widgets/home_post_rank_item.dart';
+import 'package:ondo/presentation/home/widgets/home_post_rank_list.dart';
 import 'package:ondo/presentation/home/widgets/home_profile_card.dart';
+import 'package:ondo/presentation/home/widgets/home_recommend_chat_list.dart';
+import 'package:ondo/presentation/home/widgets/home_recommend_post_list.dart';
 
 import '../../../core/design_system/components/post/base_post_list.dart';
 import '../../../core/design_system/components/post/post_item.dart';
@@ -51,7 +54,7 @@ class HomeScreen extends StatelessWidget {
         padding: AppPadding.screenHorizontal,
         child: Column(
           children: [
-            _PostRankList(),
+            HomePostRankList(),
             AppGap.v16,
           ],
         ),
@@ -65,201 +68,12 @@ class HomeScreen extends StatelessWidget {
       child: Column(
         children: [
           AppGap.v16,
-          _RecommendChatList(),
+          HomeRecommendChatList(),
           AppGap.v16,
-          _RecommendPostList(),
+          HomeRecommendPostList(),
           AppGap.v16,
         ],
       ),
     );
-  }
-}
-
-@immutable
-class _PostRankList extends StatefulWidget {
-  final List<Map<String, dynamic>> posts;
-
-  _PostRankList()
-    : posts = [
-        for (int i = 0; i < 5; i++) ...{{}, {}, {}, {}},
-      ];
-
-  @override
-  State<_PostRankList> createState() => _PostRankListState();
-}
-
-class _PostRankListState extends State<_PostRankList> {
-  late final PageController _pageController;
-
-  ValueNotifier<int> curIndex = ValueNotifier(0);
-
-  @override
-  void initState() {
-    _pageController = PageController(initialPage: curIndex.value);
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _pageController.dispose();
-    super.dispose();
-  }
-
-  int _getPageTotal() => (widget.posts.length / 3).ceil();
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 220,
-      width: double.infinity,
-      decoration: BoxDecoration(color: AppColors.white),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            "실시간 인기 게시물",
-            style: AppTextStyles.titleBold16(textColor: AppColors.gray90),
-          ),
-          AppGap.v16,
-          Expanded(child: _postList()),
-          AppGap.v16,
-          _indicator(),
-        ],
-      ),
-    );
-  }
-
-  Widget _postList() {
-    return PageView.builder(
-      controller: _pageController,
-      onPageChanged: (value) {
-        curIndex.value = value;
-      },
-      scrollDirection: Axis.horizontal,
-      itemBuilder: (context, pageIndex) {
-        return Column(
-          spacing: AppSpacing.s16,
-          children: List.generate(3, (itemIndex) {
-            final currentItemIndex = (pageIndex * 3) + itemIndex;
-
-            return currentItemIndex < widget.posts.length
-                ? HomePostRankItem(
-                    title: "요즘 공부 어케 하시나요 다들",
-                    createAgo: 3,
-                    favorite: 160,
-                    rank: currentItemIndex + 1,
-                  )
-                : SizedBox.shrink();
-          }),
-        );
-      },
-      itemCount: _getPageTotal(),
-    );
-  }
-
-  Widget _indicator() {
-    return ValueListenableBuilder(
-      valueListenable: curIndex,
-      builder: (context, value, _) {
-        return Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          spacing: AppSpacing.s6,
-          children: List.generate(
-            3,
-            (index) => _indicatorIcon(index == value % 3),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _indicatorIcon(bool isFocus) {
-    return Container(
-      width: AppSpacing.s4,
-      height: AppSpacing.s4,
-      decoration: BoxDecoration(
-        color: isFocus ? AppColors.gray80 : AppColors.gray60,
-        borderRadius: AppRadius.circleRadius,
-      ),
-    );
-  }
-}
-
-@immutable
-class _RecommendChatList extends StatelessWidget {
-  final List<Map<String, dynamic>> _chats;
-
-  _RecommendChatList()
-    : _chats = [
-        for (int i = 0; i < 5; i++) ...{{}, {}, {}, {}},
-      ];
-
-  int _getPageTotal() => (_chats.length / 3).ceil();
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 200,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            "커피챗 추천",
-            style: AppTextStyles.titleBold16(),
-          ),
-          AppGap.v16,
-          Expanded(child: _chatList()),
-        ],
-      ),
-    );
-  }
-
-  Widget _chatList() {
-    return PageView.builder(
-      scrollDirection: Axis.horizontal,
-      itemBuilder: (_, pageIndex) {
-        return Row(
-          spacing: AppSpacing.s16,
-          children: List.generate(3, (itemIndex) {
-            final int currentItemIndex = (pageIndex * 3) + itemIndex;
-
-            return currentItemIndex < _chats.length
-                ? Flexible(
-                    child: HomeProfileCard(
-                      skill: "UI/UX",
-                      name: "김유찬",
-                      rating: 4,
-                    ),
-                  )
-                : SizedBox.shrink();
-          }),
-        );
-      },
-      itemCount: _getPageTotal(),
-    );
-  }
-}
-
-@immutable
-class _RecommendPostList extends BasePostGrid {
-  final List<Map<String, dynamic>> posts;
-
-  _RecommendPostList({super.title = "추천 게시물"})
-    : posts = [
-        for (int i = 0; i < 5; i++) ...{{}, {}, {}},
-      ];
-
-  @override
-  List<Widget> listBuilder() {
-    return List.generate(posts.length, (index) {
-      return PostItem(
-        skills: ["UI/UX", "FrontEnd"],
-        title: "요즘 UI UX",
-        author: "김유찬",
-        bookmarks: 12,
-        favorites: 12,
-        createMinutes: 4,
-      );
-    });
   }
 }
